autoops/services/model_service/requirements.txt
# AutoOps Model Service Requirements
# Core dependencies for the model serving service

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Monitoring and metrics
prometheus-client==0.19.0

# Data validation
pydantic==2.5.0

# HTTP client
httpx==0.25.2

# Logging
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
ruff==0.1.6
isort==5.12.0

# Deep Learning and AI libraries
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
datasets==2.14.0

# Additional ML libraries
xgboost==2.0.2
lightgbm==4.1.0

# AI/ML utilities
accelerate==0.24.0
tokenizers==0.15.0
sentencepiece==0.1.99

# LLM and NLP utilities
sentence-transformers==2.2.2
optimum==1.14.0

# API integrations
google-generativeai==0.3.2
openai==1.3.0
